+++ lib/core/constants/images.dart
@@ -1,3 +1,9 @@
 class AppImages{
   static const logo = 'assets/logo.png';
+
+  ///avatars
+  static const user1 = 'assets/avatars/user1.jpg';
+  static const user2 = 'assets/avatars/user2.jpg';
+  static const user3 = 'assets/avatars/user3.jpg';
+  static const user4 = 'assets/avatars/user4.jpg';
 }